功能完善：对齐学术文档并完善战略关系系统
核心改动：
- 增强智能体决策提示词模板，明确区分system和user prompt
- 完善国家战略关系属性系统，支持关系配对规则
- 前端增加战略关系配置界面和可视化
- 新增测试脚本和工具文件

+++ app/api/action_config.py
@@ -24,8 +24,8 @@ class ActionConfigResponse(BaseModel):
     action_category: str
     action_desc: str
     respect_sov: bool
-    initiator_power_change: int
-    target_power_change: int
+    initiator_power_change: float
+    target_power_change: float
     is_initiative: bool
     is_response: bool
